CP-IJ: Refactor to server-based model
This works around the blocked PIPE communication between subprocesses.
