+++ cpij/__init__.py
@@ -1,8 +1,4 @@
 # -*- coding: utf-8 -*-
-"""
-CellStar package providing CellStar algorithm for segmentation of yeast cells in brightfield imagery.
-Date: 2013-2016
-Website: http://cellstar-algorithm.org/
-"""
-__author__ = 'Mark Hiner'
-__all__ = ["bridge"]
+
+__author__ = 'Mark Hiner, Alice Lucas, Beth Cimini'
+__all__ = ["bridge", "server"]
